+++ R/plot.gcfit.R
@@ -2,7 +2,7 @@
 #'
 #' @param x A fit for some growth data
 #' @param y Not used
-#' @param show_curve Whether or not to show the fitted curve (default: \code{TRUE})
+#' @param show_fit Whether or not to show the fitted curve (default: \code{TRUE})
 #' @param show_data Whether or not to show the original data (default \code{TRUE})
 #' @param show_maxrate Whether or not to show a tangent line where the maximum growth rate occurs (default \code{TRUE})
 #' @param show_asymptote Whether or not to indicate the maximum growth level (default \code{FALSE})
@@ -17,7 +17,7 @@
 #' lfit <- fit_growth_logistic(df=mydata, Time, OD600)
 #' plot(lfit)}
 #' 
-plot.gcfit <- function(x, y = NULL, show_curve = TRUE, show_data = TRUE,
+plot.gcfit <- function(x, y = NULL, show_fit = TRUE, show_data = TRUE,
                        show_maxrate = TRUE, show_asymptote = FALSE,
                        show_lag = FALSE, ...) {
     opt_args <- list(...)
@@ -35,7 +35,7 @@ plot.gcfit <- function(x, y = NULL, show_curve = TRUE, show_data = TRUE,
     opt_args$show_asymptote <- NULL
 
     # Plot the fitted curve
-    if (show_curve) {
+    if (show_fit) {
         try(do.call(plot, opt_args))
     }