+++ NEWS.md
@@ -1,7 +1,7 @@
 # growthcurve 0.1.4
 
 - Renamed 'raw' to 'data' in gfcit objects
-- Added `autoplot.gcfit` function
+- Added `autoplot.gcfit` function and alias `gggrowthcurve`
 - Added [broom](https://github.com/dgrtwo/broom)-compatible `tidy` function
 
 # growthcurve 0.1.3